+++ packages/xchain-ethereum/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './etherscan-api-types'
+export * from './ethplorer-api-types'
 export * from './client-types'